Separates filter panels from dashboard panels
